refactor(frontend): App.jsxをさらに分割し1,000行以下にする（issue #607）
設計:
- issue #607のクローズ基準「App.jsxが1,000行以下」達成に向け、残る
  大きめの塊をさらに専用view/componentへ切り出した
- views/DivisionSelectView.jsx: division未選択時の「どなた向けに遊ぶか」
  選択画面（開設中クイズ大会ルーム一覧・各種リンクを含む）
- views/CategorySelectView.jsx: division選択後のかるた種類選択画面
  （絵札印刷上限バリデーション表示・決定/印刷ボタンを含む）
- components/SettingsFooter.jsx: テーマ/言語/声/順番/スピード/回数の
  設定UI（プレーンな制御コンポーネント、業務ロジックを持たない）
- components/QuizRoomActionsPanel.jsx: クイズ大会モードのルーム作成/
  再開/接続状態表示/早押し判定モーダルの表示切り替え
- components/PlayerRegistrationPanel.jsx: 「取った人を記録する」参加者
  登録UI（QuizRoomActionsPanelから利用）
- hooks/usePlayerScores.js: 参加者・スコア管理（issue #518）一式を
  切り出した。players/scoresByCategoryのstate、currentScores/
  scoreSummaryのメモ、ラウンド切り替え検知（レンダー中state調整）、
  addPlayer/removePlayer/recordTaken、restartCategoryから呼ばれる
  resetCategoryScores()を含む

影響:
- frontend/src/App.jsx: 1,301行→989行（issue #607のクローズ基準達成）
- 新規ファイル6件（views 2件、components 3件、hooks 1件）
- 各画面・設定UI・参加者登録・スコア集計の挙動に機能的な変更はない
  （プレゼンテーション層の分割のみ）

テスト:
- frontend: npx eslint . エラー0件（プロジェクト全体）
- frontend: npx vitest run src/App.test.jsx を各段階の変更ごとに実行し、
  65件全て成功することを確認（division/category分割時点、settings/
  quiz-room panel分割時点、usePlayerScores分割時点の3回）
- backend: npx vitest run 1543件全て成功、npx eslint . エラー0件
  （今回未変更だが対象パッケージとして確認）

issue #607はクローズ基準達成のためクローズ対応
